Rename patch2 to patch
