+++ pkg/patch/utils_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package patch2
+package patch
 
 import (
 	"testing"
