+++ pkg/prober/context/probe_context.go
@@ -9,8 +9,7 @@ import (
 	"github.com/go-logr/logr"
 
 	vmopv1 "github.com/vmware-tanzu/vm-operator/api/v1alpha2"
-
-	patch "github.com/vmware-tanzu/vm-operator/pkg/patch2"
+	"github.com/vmware-tanzu/vm-operator/pkg/patch"
 )
 
 // ProbeContext is the context used for VM Probes.
